Work in progress

+++ include/fejix/core/vec.h
@@ -10,10 +10,10 @@ struct fj_vec {
     void */*[]?*/ items;
 
     /* Number of currently stored elements */
-    uint32_t length;
+    size_t length;
 
     /* Number of currently allocated elements */
-    uint32_t capacity;
+    size_t capacity;
 
     size_t item_size;
 };
@@ -26,25 +26,25 @@ void fj_vec_deinit(struct fj_vec * vec);
 
 
 /** If the new capacity is smaller that the current length, the length is trancated. */
-fj_err_t fj_vec_resize(struct fj_vec * vec, uint32_t capacity);
+fj_err_t fj_vec_resize(struct fj_vec * vec, size_t capacity);
 
 /** Ensures that the vector has at least the reserved elements.
     That is, the capacity is at least `length + reserved_items` */
-fj_err_t fj_vec_resize_to_reserve(struct fj_vec * vec, uint32_t reserved_items);
+fj_err_t fj_vec_resize_to_reserve(struct fj_vec * vec, size_t reserved_items);
 
 /** Sets the capacity to the vector length. */
 fj_err_t fj_vec_resize_to_fit(struct fj_vec * vec);
 
 
-uint32_t fj_vec_get_last_index(struct fj_vec const * vec);
+size_t fj_vec_get_last_index(struct fj_vec const * vec);
 
-uint32_t fj_vec_get_push_index(struct fj_vec const * vec);
+size_t fj_vec_get_push_index(struct fj_vec const * vec);
 
 fj_bool32_t fj_vec_is_empty(struct fj_vec const * vec);
 
 fj_bool32_t fj_vec_has_allocated(struct fj_vec const * vec);
 
-void */*[]?*/ fj_vec_offset(struct fj_vec const * vec, uint32_t offset_index);
+void */*[]?*/ fj_vec_offset(struct fj_vec const * vec, size_t offset_index);
 
 
 /** Copies the items into the list.
@@ -53,25 +53,25 @@ void */*[]?*/ fj_vec_offset(struct fj_vec const * vec, uint32_t offset_index);
 void fj_vec_replace_items(
     struct fj_vec * vec,
     void const * items,
-    uint32_t destination_index,
-    uint32_t item_count
+    size_t destination_index,
+    size_t item_count
 );
 
 fj_err_t fj_vec_insert_uninit(
     struct fj_vec * vec,
-    uint32_t destination_index,
-    uint32_t item_count
+    size_t destination_index,
+    size_t item_count
 );
 
 fj_err_t fj_vec_insert_items(
     struct fj_vec * vec,
     void const * items,
-    uint32_t destination_index,
-    uint32_t item_count
+    size_t destination_index,
+    size_t item_count
 );
 
 /** Clamps `start_index` and `start_index+item_count` to `0..length`. */
-fj_err_t fj_vec_remove_items(struct fj_vec * vec, uint32_t start_index, uint32_t item_count);
+fj_err_t fj_vec_remove_items(struct fj_vec * vec, size_t start_index, size_t item_count);
 
 /** Works like `insert`. Copies the item into the end of the list. */
 fj_err_t fj_vec_push_item(struct fj_vec * vec, void const * item);
