+++ src/core/vec.c
@@ -1,4 +1,3 @@
-#include "fejix/core/error.h"
 #include <fejix/core/vec.h>
 
 #include <fejix/core/malloc.h>
@@ -7,12 +6,12 @@
 #include <string.h>
 
 
-uint32_t fj_vec_get_last_index(struct fj_vec const * vec)
+size_t fj_vec_get_last_index(struct fj_vec const * vec)
 {
     return vec->length - 1;
 }
 
-uint32_t fj_vec_get_push_index(struct fj_vec const * vec)
+size_t fj_vec_get_push_index(struct fj_vec const * vec)
 {
     return vec->length;
 }
@@ -27,7 +26,7 @@ fj_bool32_t fj_vec_has_allocated(struct fj_vec const * vec)
     return vec->items != NULL;
 }
 
-void */*[]?*/ fj_vec_offset(struct fj_vec const * vec, uint32_t offset_index)
+void */*[]?*/ fj_vec_offset(struct fj_vec const * vec, size_t offset_index)
 {
     if (offset_index >= vec->capacity) {
         return NULL;
@@ -39,7 +38,7 @@ void */*[]?*/ fj_vec_offset(struct fj_vec const * vec, uint32_t offset_index)
 
 /** Small optimization to reduce the amount of allocations. */
 static
-uint32_t vec_get_min_capacity(struct fj_vec * vec)
+size_t vec_get_min_capacity(struct fj_vec * vec)
 {
     if (vec->item_size == 1) {
         return 8;
@@ -52,9 +51,9 @@ uint32_t vec_get_min_capacity(struct fj_vec * vec)
 
 
 static
-uint32_t vec_get_capacity_to_grow(struct fj_vec * vec, uint32_t grow_count)
+size_t vec_get_capacity_to_grow(struct fj_vec * vec, size_t grow_count)
 {
-    uint32_t new_capacity = FJ_MAX(vec->capacity, 1);
+    size_t new_capacity = FJ_MAX(vec->capacity, 1);
 
     while (vec->length + grow_count > new_capacity) {
         new_capacity *= 2;
@@ -65,9 +64,9 @@ uint32_t vec_get_capacity_to_grow(struct fj_vec * vec, uint32_t grow_count)
 
 
 static
-uint32_t vec_get_capacity_to_shrink(struct fj_vec * vec)
+size_t vec_get_capacity_to_shrink(struct fj_vec * vec)
 {
-    uint32_t new_capacity = vec->capacity;
+    size_t new_capacity = vec->capacity;
 
     while (vec->length < new_capacity/4) {
         new_capacity /= 2;
@@ -77,7 +76,7 @@ uint32_t vec_get_capacity_to_shrink(struct fj_vec * vec)
 }
 
 
-fj_err_t fj_vec_resize(struct fj_vec * vec, uint32_t capacity)
+fj_err_t fj_vec_resize(struct fj_vec * vec, size_t capacity)
 {
     FJ_WITH_ERRORS
 
@@ -96,9 +95,9 @@ fj_err_t fj_vec_resize(struct fj_vec * vec, uint32_t capacity)
 }
 
 
-fj_err_t fj_vec_resize_to_reserve(struct fj_vec * vec, uint32_t reserved_items)
+fj_err_t fj_vec_resize_to_reserve(struct fj_vec * vec, size_t reserved_items)
 {
-    uint32_t new_capacity = vec_get_capacity_to_grow(vec, reserved_items);
+    size_t new_capacity = vec_get_capacity_to_grow(vec, reserved_items);
     return fj_vec_resize(vec, new_capacity);
 }
 
@@ -112,27 +111,27 @@ fj_err_t fj_vec_resize_to_fit(struct fj_vec * vec)
 static
 fj_err_t vec_maybe_shrink(struct fj_vec * vec)
 {
-    uint32_t new_capacity = vec_get_capacity_to_shrink(vec);
+    size_t new_capacity = vec_get_capacity_to_shrink(vec);
     return fj_vec_resize(vec, new_capacity);
 }
 
 
 static
-void shift_items_for_insert(struct fj_vec * vec, uint32_t source_index, uint32_t item_distance)
+void shift_items_for_insert(struct fj_vec * vec, size_t source_index, size_t item_distance)
 {
     uint8_t * src = fj_vec_offset(vec, source_index);
     uint8_t * dst = src + item_distance * vec->item_size;
-    uint32_t item_move_count = vec->length - source_index;
+    size_t item_move_count = vec->length - source_index;
     memmove(dst, src, item_move_count * vec->item_size);
 }
 
 
 static
-void shift_items_for_remove(struct fj_vec * vec, uint32_t source_index, uint32_t item_distance)
+void shift_items_for_remove(struct fj_vec * vec, size_t source_index, size_t item_distance)
 {
     uint8_t * src = fj_vec_offset(vec, source_index);
     uint8_t * dst = src - item_distance * vec->item_size;
-    uint32_t item_move_count = vec->length - source_index;
+    size_t item_move_count = vec->length - source_index;
     memmove(dst, src, item_move_count * vec->item_size);
 }
 
@@ -159,8 +158,8 @@ void fj_vec_deinit(struct fj_vec * vec)
 void fj_vec_replace_items(
     struct fj_vec * vec,
     void const * items,
-    uint32_t destination_index,
-    uint32_t item_count
+    size_t destination_index,
+    size_t item_count
 )
 {
     uint8_t * destination = fj_vec_offset(vec, destination_index);
@@ -169,7 +168,7 @@ void fj_vec_replace_items(
 }
 
 
-fj_err_t fj_vec_insert_uninit(struct fj_vec * vec, uint32_t destination_index, uint32_t item_count)
+fj_err_t fj_vec_insert_uninit(struct fj_vec * vec, size_t destination_index, size_t item_count)
 {
     FJ_WITH_ERRORS
 
@@ -190,8 +189,8 @@ fj_err_t fj_vec_insert_uninit(struct fj_vec * vec, uint32_t destination_index, u
 fj_err_t fj_vec_insert_items(
     struct fj_vec * vec,
     void const * items,
-    uint32_t destination_index,
-    uint32_t item_count
+    size_t destination_index,
+    size_t item_count
 )
 {
     FJ_WITH_ERRORS
@@ -206,7 +205,7 @@ fj_err_t fj_vec_insert_items(
 }
 
 
-fj_err_t fj_vec_remove_items(struct fj_vec * vec, uint32_t start_index, uint32_t item_count)
+fj_err_t fj_vec_remove_items(struct fj_vec * vec, size_t start_index, size_t item_count)
 {
     if (start_index + item_count <= fj_vec_get_last_index(vec)) {
         shift_items_for_remove(vec, start_index+item_count, item_count);
